+++ src/simulate/heap.cpp
@@ -165,8 +165,8 @@ namespace das {
     void PersistentStringAllocator::forEachString ( const callable<void (const char *)> & fn ) {
         for ( uint32_t si=0; si!=DAS_MAX_SHOE_CUNKS; ++si ) {
             for ( auto ch=model.shoe.chunks[si]; ch; ch=ch->next ) {
-                uint32_t utotal = ch->total / 32;
-                for ( uint32_t i=0; i!=utotal; ++i ) {
+                uint64_t utotal = ch->total / 32;
+                for ( uint64_t i=0; i!=utotal; ++i ) {
                     uint32_t b = ch->bits[i];
                     for ( uint32_t j=0; j!=32; ++j ) {    // todo: simpler bit loop
                         if ( b & (1<<j) ) {
@@ -412,8 +412,8 @@ namespace das {
                 totalChunks ++;
                 tout << "\t" << HEX << "[" << uint64_t(ch->data) << ".." << (uint64_t(ch->data)+(ch->size*ch->total)) << ")\n" << DEC;
                 tout << "\t" << ch->allocated << " of " << ch->total << ", " << (ch->allocated*ch->size) << " of " << ch->totalBytes << " bytes\n";
-                uint32_t utotal = ch->total / 32;
-                for ( uint32_t i=0; i!=utotal; ++i ) {
+                uint64_t utotal = ch->total / 32;
+                for ( uint64_t i=0; i!=utotal; ++i ) {
                     uint32_t b = ch->bits[i];
                     for ( uint32_t j=0; j!=32; ++j ) {
                         if ( b & (1<<j) ) {
